changed api logic and fixed with states on search/searchbar
