+++ src/types/my-theme.d.ts
